Reformat code
You don't need to reformat code for its own sake, but rather when
making changes it's good practice to check in a consistent style.  In
this case I was going to track down one of the seven style warnings,
but ran out of time, and see no reason not to check in the reformatted
code.